requirements.txt
requests==2.31.0
pandas==2.1.3
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.24.3
beautifulsoup4==4.12.2
aiohttp==3.9.1
python-dotenv==1.0.0
schedule==1.2.0